migrate missing changes from PR #5799
